Knight can find the shortest path from a start to an end
Renamed the position function to chessSquare added a "parent" variable which stores which square the knight moved from to this square.
Creates a getShortestKnightPath that uses BFS to reach the end position, once this is reached it will read each parent square and add them to an array which is finally returned.